Create A8RampsConfig.h
This is my Configuration.h for Marlin 1.1.x for the Anet A8 with RAMPS 1.4 upgrade, DRV8825 Stepper Drivers with 1/32 microstepping and bed levelling with an induction sensor. This is a fairly common setup and when I originally set it up I had a hard time finding the right info for calculating the proper steps/mm for my microstepping along with other options. This will work for you with a similar setup, though will need to make changes to your bed sensor position as I am using an E3D volcano with a custom carriage & mount that isn't stock.